Заканчивает задание
