+++ src/utils/point.js
@@ -3,15 +3,6 @@ import utc from 'dayjs/plugin/utc';
 
 dayjs.extend(utc);
 
-const getRandomInteger = (a = 0, b = 1) => {
-  const lower = Math.ceil(Math.min(a, b));
-  const upper = Math.floor(Math.max(a, b));
-
-  return Math.floor(lower + Math.random() * (upper - lower + 1));
-};
-
-const getRandomArrayElement = (items) => items[getRandomInteger(0, items.length - 1)];
-
 const correctDateFormat = (date) => dayjs.utc(date).format('DD/MM/YY_HH:mm');
 
 const extractDate = (date) => dayjs.utc(date).format('MMM DD');
@@ -32,4 +23,4 @@ const calculateFlightTime = (startTime, endTime, unit = 'm') =>{
   }
 };
 
-export { getRandomArrayElement, getRandomInteger, correctDateFormat, extractDate, extractTime, calculateFlightTime };
+export { correctDateFormat, extractDate, extractTime, calculateFlightTime };
