+++ src/view/board-view.js
@@ -1,4 +1,3 @@
-//import { createElement } from '../render.js';
 import AbstractView from '../framework/view/abstract-view.js';
 
 const createBoardTemplate = () => '<section class="terip-vents"></section>';
