+++ src/view/filters-view.js
@@ -1,4 +1,3 @@
-//import { createElement } from '../render.js';
 import AbstractView from '../framework/view/abstract-view.js';
 
 const createFiltersTemplate = () => (
